Move / rename CollectionIterator to IndexedIterator
